update-1

+++ bch.js
@@ -7,12 +7,14 @@ async function start() {
         const mnemonic = JSON.parse(localStorage.getItem("wallet")).mnemonic;
         const wallet = await TestNetSmartBchWallet.fromSeed(mnemonic);
         showData(wallet);
-        console.log(wallet);
+        // console.log(wallet);
         // myWallet(wallet);
         // qr(wallet);
         document.getElementById("mainBox").style.display = "none";
     } else {
         document.getElementById("myWalletBox").style.display = "none";
+        document.getElementById("accountBox").style.display = "none";
+        document.getElementById("mainBox").style.display = "flex";
     }
 }
 
@@ -20,7 +22,7 @@ async function newWallet() {
     localStorage.clear("wallet");
     const wallet = await TestNetSmartBchWallet.newRandom();
     localStorage.setItem("wallet", JSON.stringify(wallet));
-    console.log(wallet);
+    // console.log(wallet);
     showData(wallet);
     document.getElementById("mainBox").style.display = "none";
     document.getElementById("myWalletBox").style.display = "flex";
@@ -34,19 +36,28 @@ async function showData(wallet) {
     document.getElementById("walletBalance").innerText = `${balance.bch.toFixed(8)} BCH`;
 
     document.getElementById("walletAddress").value = `${wallet.address}`;
+    document.getElementById("walletAddress2").value = `${wallet.address}`;
 
     document.querySelector('#deposit').src = wallet.getDepositQr().src;
+    document.getElementById("showMnemonic").innerText = `${wallet.mnemonic}`;
+    // document.getElementById("showKey").innerText = `${wallet.privateKey}`;
 }
 
 async function importWallet() {
-    localStorage.clear("wallet");
-    let mnemonic = document.getElementById("seed").value;
-    const wallet = await TestNetSmartBchWallet.fromSeed(mnemonic);
-    localStorage.setItem("wallet", JSON.stringify(wallet));
-    showData(wallet);
-    console.log(wallet);
-    document.getElementById("mainBox").style.display = "none";
-    document.getElementById("myWalletBox").style.display = "flex";
+    if (document.getElementById("seed").value != "") {
+        localStorage.clear("wallet");
+        let mnemonic = document.getElementById("seed").value;
+        document.getElementById("seed").value = "";
+        const wallet = await TestNetSmartBchWallet.fromSeed(mnemonic);
+        localStorage.setItem("wallet", JSON.stringify(wallet));
+        showData(wallet);
+        // console.log(wallet);
+        document.getElementById("mainBox").style.display = "none";
+        document.getElementById("myWalletBox").style.display = "flex";
+    } else {
+        console.log('invalid mnemonic');
+    }
+
 }
 
 function closeSend() {
@@ -73,9 +84,8 @@ function openReceive() {
 
 function copyAddress() {
     let input = document.getElementById("walletAddress");
-    console.log(input);
-    // input.focus();
-    input.select();
+    // console.log(input);
+    // console.log(input.value);
     navigator.clipboard.writeText(input.value);
 }
 
@@ -88,12 +98,12 @@ async function withdraw() {
         const wallet = await TestNetSmartBchWallet.fromSeed(mnemonic);
         // console.log(wallet);
         const balance = await wallet.getBalance();
-        const overrides = {gasPrice: 0.2,gasLimit:26038}
+        const overrides = { gasPrice: 0.2, gasLimit: 26038 }
         const txData = await wallet.send([[seller, amount, 'bch', overrides]]);
         console.log(txData);
 
         localStorage.setItem("txData", JSON.stringify(txData));
-        txFees(balance,wallet,seller,txData,amount);
+        txFees(balance, wallet, seller, txData, amount);
         showData(wallet);
 
     } else {
@@ -103,6 +113,7 @@ async function withdraw() {
 
 //withdraw section -->
 
+// ---Ignore this---
 // function test() {
 //     let txData =JSON.parse(localStorage.getItem("txData"));
 //     console.log(txData[0]);
@@ -128,7 +139,7 @@ async function maxAmount() {
     const max = await wallet.getMaxAmountToSend();
     console.log(max);
     const maxbch = max["bch"];
-    document.getElementById("amount").value =`${maxbch}`;
+    document.getElementById("amount").value = `${maxbch}`;
 }
 
 function closeAccount() {
@@ -149,7 +160,7 @@ function transaction(wallet) {
     document.getElementById("menuBox").style.display = "none";
 }
 
-async function txFees(balance,wallet,seller,txData,amount) {
+async function txFees(balance, wallet, seller, txData, amount) {
     console.log(balance.bch);
     const totalSend = balance.bch - txData[0].balance["bch"];
     console.log(`TotalSend: ${totalSend.toFixed(8)}`);
@@ -161,3 +172,42 @@ async function txFees(balance,wallet,seller,txData,amount) {
     console.log(`TxId: ${txData[0].txId}`);
     console.log(`BalanceLeft: ${txData[0].balance["bch"].toFixed(8)}`);
 }
+
+function resetFunction() {
+    localStorage.clear("wallet");
+    start();
+}
+
+// function copyKey() {
+//     let key = document.getElementById("showKey");
+//     navigator.clipboard.writeText(key.innerText);
+// }
+
+function copyMnemonic() {
+    let mnemonic = document.getElementById("showMnemonic");
+    navigator.clipboard.writeText(mnemonic.innerText);
+}
+
+// ---Ignore this---
+// const createToken = async ()=>{
+//     const mnemonic = JSON.parse(localStorage.getItem("wallet")).mnemonic;
+//     const wallet = await TestNetSmartBchWallet.fromSeed(mnemonic);
+
+//    const tokenName = document.getElementById("tokenName").value;
+//    const tokenTicker = document.getElementById("tokenTicker").value;
+//    const tokenDecimals = document.getElementById("tokenDecimals").value;
+//    const tokenQuantity = document.getElementById("tokenQuantity").value;
+//    const endMint = document.getElementById("endMint").value;
+
+//     const genesisOptions = {
+//         name: tokenName,
+//         ticker: tokenTicker,
+//         decimals: Number(tokenDecimals),
+//         initialAmount: Number(tokenQuantity),
+//         endBaton: Boolean(endMint),
+//         endBaton: true,
+//       };
+//       const {tokenId} = await wallet.sep20.genesis(genesisOptions);
+//       console.log({tokenId});
+//       console.log(tokenId);
+// }
